+++ src/packages/piral-cli-parcel/src/parcel/bundler.ts
@@ -235,7 +235,17 @@ export async function postProcess(bundle: Bundler.ParcelBundle, version: PiletSc
                 `e.href=${bundleUrlRef}+${JSON.stringify(cssName)}`,
                 `d.head.appendChild(e)`,
               ].join(';');
-              result = `(function(){${stylesheet}})();${result}`;
+
+              /**
+               * Only happens in debug mode:
+               * Apply this only when the stylesheet is not yet part of the file.
+               * This solves the edge case of touching files (i.e., saving without any change).
+               * Here, Parcel triggers a re-build, but does not change the output files.
+               * Making the change here would destroy the file.
+               */
+              if (result.indexOf(stylesheet) === -1) {
+                result = `(function(){${stylesheet}})();${result}`;
+              }
             }
 
             // Only happens in (pilet) debug mode: